+++ addons_updater/CHANGELOG.md
@@ -1,4 +1,6 @@
 - The Home Assistant project has deprecated support for the armv7, armhf and i386 architectures. Support wil be fully dropped in the upcoming Home Assistant 2025.12 release
+- Add support for codeberg source handling and increment version
+
 
 ## 3.19.12 (18-11-2025)